Final batch of optimizations for this time around.
We could think about an algorithmic optimization for move (and copy) to avoid a second walk of the tree to find the source value, if it shows up in the tree earlier, but that would make "small-ish" documents less performant.

The goal needs to be "good performance in a wide range of real-world scenarios" and smallish (but not quite as small as the current test data) are probably more realistic payloads to work on.