Avoid string allocation with Duration parsing.
Unfortunately, NodaTime.Period is a class, not a readonly struct, so it unncessarily allocates at that stage. I've created a Corvus.Json.Period which is similar, but a readonly struct, but can be implcitly cast to/from NodaTime.Period for interoperability with the rest of the NodaTime world, but for basic usage avoids allocating.