Removed [NotNull] from the two not-nullable parameters on ExecuteAsync. These came from the Spectre sample code, and are not necessary.
Throwing if SchemaFile or RootNamespace are null; even though the Spectre framework guarantees this for us, the compiler doesn't know this. @idg10 commented that this was confusing.

In retrospect, I agree, because although it expresses the reality of the situation, it does so in a way which is not common; it is easier to blow up explicitly at runtime to express the same problem.