Added optimizations for  - property name equality checks - aggressive inlining
These (broadly) compensate for the fact that we have fixed a number of cases where we were not pushing the validation property location correctly to the context in V2. Those extra instructions (even though we don't push the location itself in the flags case, just test to see if we should) add a small overhead.