Fixed naming heuristics - Added a proper way of systematically weeding out multiple paths to the *same* type, via dynamic references.
This situation occurs when there are multiple references to a subschema that all actually occur in the same dynamic context, but this cannot be determined until you are "done" (you can tell because the dynamic types will all have the same name, end up in the same parent, because that parent represents the dynamic anchor for the type. If they are from a different dynamic scope, they will have a different parent.

V2 deals with this by the brute-force method of just writing the files over and over again. This is a wasteful (especially in the Source Generator scenario) so we avoid that in V3.